+++ deepchem/models/tensorgraph/models/graph_models.py
@@ -1,23 +1,21 @@
+import math
+
 import numpy as np
 import six
 import tensorflow as tf
-
+from deepchem.data import NumpyDataset
+from deepchem.feat.graph_features import ConvMolFeaturizer
 from deepchem.feat.mol_graphs import ConvMol
-from deepchem.metrics import to_one_hot, from_one_hot
+from deepchem.metrics import to_one_hot
 from deepchem.models.tensorgraph.graph_layers import WeaveLayer, WeaveGather, \
   Combine_AP, Separate_AP, DTNNEmbedding, DTNNStep, DTNNGather, DAGLayer, \
   DAGGather, DTNNExtract, MessagePassing, SetGather
 from deepchem.models.tensorgraph.layers import Dense, Concat, SoftMax, \
   SoftMaxCrossEntropy, GraphConv, BatchNorm, \
-  GraphPool, GraphGather, WeightedError, Dropout, BatchNormalization, Stack, Layer, Flatten
+  GraphPool, GraphGather, WeightedError, Dropout, BatchNormalization, Stack, Layer, Flatten, GraphCNNLayer, GraphCNNPool
 from deepchem.models.tensorgraph.layers import L2Loss, Label, Weights, Feature
 from deepchem.models.tensorgraph.tensor_graph import TensorGraph
 from deepchem.trans import undo_transforms
-from deepchem.utils.evaluate import GeneratorEvaluator
-from deepchem.data import NumpyDataset
-from deepchem.data.data_loader import featurize_smiles_np
-from deepchem.feat.graph_features import ConvMolFeaturizer
-import math
 
 
 class WeaveTensorGraph(TensorGraph):
@@ -485,197 +483,6 @@ class DAGTensorGraph(TensorGraph):
         yield feed_dict
 
 
-def make_variable(name, shape, initializer=tf.truncated_normal_initializer(), regularizer=None):
-  dtype = tf.float32
-  var = tf.get_variable(name, shape, initializer=initializer, regularizer=regularizer, dtype=dtype)
-  return var
-
-
-def make_bias_variable(name, shape):
-  dtype = tf.float32
-  var = tf.get_variable(name, shape, initializer=tf.constant_initializer(0.1), dtype=dtype)
-  return var
-
-
-def make_variable_with_weight_decay(name, shape, stddev=0.01, wd=0.0005):
-  dtype = tf.float32
-  regularizer = None
-  if wd is not None and wd > 1e-7:
-    def regularizer(var):
-      return tf.multiply(tf.nn.l2_loss(var), wd, name='weight_loss')
-  var = make_variable(name, shape, initializer=tf.truncated_normal_initializer(stddev=stddev), regularizer=regularizer)
-  return var
-
-
-def batch_mat_mult(A, B):
-  A_shape = tf.shape(A)
-  A_reshape = tf.reshape(A, [-1, A_shape[-1]])
-
-  # So the Tensor has known dimensions
-  if B.get_shape()[1] == None:
-    axis_2 = -1
-  else:
-    axis_2 = B.get_shape()[1]
-  result = tf.matmul(A_reshape, B)
-  result = tf.reshape(result, tf.stack([A_shape[0], A_shape[1], axis_2]))
-  return result
-
-
-def make_bn(input, phase, axis=-1, epsilon=0.001, mask=None, num_updates=None, name=None):
-  """
-  TODO(LESWING) hook up to layers and existing feed dick
-  Parameters
-  ----------
-  input
-  phase
-  axis
-  epsilon
-  mask
-  num_updates
-  name
-
-  Returns
-  -------
-
-  """
-  default_decay = 0.9
-  input_size = input.get_shape()[axis].value
-  if axis == -1:
-    axis = len(input.get_shape()) - 1
-  axis_arr = [i for i in range(len(input.get_shape())) if i != axis]
-  if mask == None:
-    batch_mean, batch_var = tf.nn.moments(input, axis_arr)
-  else:
-    batch_mean, batch_var = tf.nn.weighted_moments(input, axis_arr, mask)
-  gamma = make_variable('%s_gamma' % name, input_size, initializer=tf.constant_initializer(1))
-  beta = make_bias_variable('%s_bias' % name, input_size)
-  ema = tf.train.ExponentialMovingAverage(decay=default_decay, num_updates=num_updates)
-
-  def mean_var_with_update():
-    ema_apply_op = ema.apply([batch_mean, batch_var])
-    with tf.control_dependencies([ema_apply_op]):
-      return tf.identity(batch_mean), tf.identity(batch_var)
-
-  # mean, var = tf.cond(phase, mean_var_with_update, lambda: (ema.average(batch_mean), ema.average(batch_var)))
-  mean, var = mean_var_with_update()
-
-  return tf.nn.batch_normalization(input, mean, var, beta, gamma, 1e-3)
-
-
-def make_embedding_layer(V, no_filters, name="default"):
-  no_features = V.get_shape()[-1].value
-  W = make_variable_with_weight_decay('%s_weights' % name, [no_features, no_filters],
-                                      stddev=1.0 / math.sqrt(no_features))
-  b = make_bias_variable('%s_bias' % name, [no_filters])
-  V_reshape = tf.reshape(V, (-1, no_features))
-  s = tf.slice(tf.shape(V), [0], [len(V.get_shape()) - 1])
-  s = tf.concat([s, tf.stack([no_filters])], 0)
-  result = tf.reshape(tf.matmul(V_reshape, W) + b, s)
-  return result
-
-
-def make_softmax_layer(V, axis=1, name=None):
-  max_value = tf.reduce_max(V, axis=axis, keep_dims=True)
-  exp = tf.exp(tf.subtract(V, max_value))
-  prob = tf.div(exp, tf.reduce_sum(exp, axis=axis, keep_dims=True))
-  return prob
-
-
-class PassThroughLayer(Layer):
-  def __init__(self, output_num, **kwargs):
-    self.output_num = output_num
-    super(PassThroughLayer, self).__init__(**kwargs)
-
-  def create_tensor(self, in_layers=None, set_tensors=True, **kwargs):
-    self.out_tensor = self.in_layers[0].out_tensors[self.output_num]
-
-
-class GraphCNNPoolLayer(Layer):
-  def __init__(self, num_vertices, **kwargs):
-    self.num_vertices = num_vertices
-    super(GraphCNNPoolLayer, self).__init__(**kwargs)
-
-  def create_tensor(self, in_layers=None, set_tensors=True, **kwargs):
-    """
-    TODO(LESWING) self.num_vertices = 1
-    Parameters
-    ----------
-    in_layers
-    set_tensors
-    kwargs
-
-    Returns
-    -------
-
-    """
-    in_tensors = self._get_input_tensors(in_layers)
-    if len(in_tensors) == 3:
-      V, A, mask = in_tensors
-    else:
-      V, A = in_tensors
-      mask = None
-    factors = make_embedding_layer(V, self.num_vertices, name='%s_Factors' % self.name)
-
-    if mask is not None:
-      factors = tf.multiply(factors, mask)
-    factors = make_softmax_layer(factors)
-
-    result = tf.matmul(factors, V, transpose_a=True)
-
-    result_A = tf.reshape(A, (tf.shape(A)[0], -1, tf.shape(A)[-1]))
-    result_A = tf.matmul(result_A, factors)
-    result_A = tf.reshape(result_A, (tf.shape(A)[0], tf.shape(A)[-1], -1))
-    result_A = tf.matmul(factors, result_A, transpose_a=True)
-    result_A = tf.reshape(result_A, (tf.shape(A)[0], self.num_vertices, A.get_shape()[2].value, self.num_vertices))
-    # We do not need the mask because every graph has self.num_vertices vertices now
-    # result = make_bn(result, True, mask=None, name="%s_bn" % self.name)
-    self.out_tensors = [result, result_A, factors]
-
-
-def GraphCNNPool(num_vertices, **kwargs):
-  gcnnpool_layer = GraphCNNPoolLayer(num_vertices, **kwargs)
-  return [PassThroughLayer(x, in_layers=gcnnpool_layer) for x in range(3)]
-
-
-class GraphCNNLayer(Layer):
-  def __init__(self, num_filters, **kwargs):
-    self.num_filters = num_filters
-    super(GraphCNNLayer, self).__init__(**kwargs)
-
-  def create_tensor(self, in_layers=None, set_tensors=True, **kwargs):
-    inputs = self._get_input_tensors(in_layers)
-    if len(inputs) == 3:
-      V, A, mask = inputs
-    else:
-      V, A = inputs
-      mask = None
-    no_A = A.get_shape()[2].value
-    no_features = V.get_shape()[2].value
-    W = make_variable_with_weight_decay('%s_weights' % self.name, [no_features * no_A, self.num_filters],
-                                        stddev=math.sqrt(1.0 / (no_features * (no_A + 1) * 1.0)))
-    W_I = make_variable_with_weight_decay('%s_weights_I' % self.name, [no_features, self.num_filters],
-                                          stddev=math.sqrt(1.0 / (no_features * (no_A + 1) * 1.0)))
-    b = make_bias_variable('%s_bias' % self.name, [self.num_filters])
-
-    n = self.graphConvolution(V, A)
-    A_shape = tf.shape(A)
-    n = tf.reshape(n, [-1, A_shape[1], no_A * no_features])
-    result = batch_mat_mult(n, W) + batch_mat_mult(V, W_I) + b
-    # result = make_bn(result, True, mask=mask, name="%s_bn" % self.name)
-    if set_tensors:
-      self.out_tensor = result
-    return result
-
-  def graphConvolution(self, V, A):
-    no_A = A.get_shape()[2].value
-    no_features = V.get_shape()[2].value
-
-    A_shape = tf.shape(A)
-    A_reshape = tf.reshape(A, tf.stack([-1, A_shape[1] * no_A, A_shape[1]]))
-    n = tf.matmul(A_reshape, V)
-    return tf.reshape(n, [-1, A_shape[1], no_A, no_features])
-
-
 class PetroskiSuchTensorGraph(TensorGraph):
   def __init__(self, n_tasks, max_atoms=200, dropout=0.2, mode="classification", **kwargs):
     """
@@ -901,9 +708,9 @@ class GraphConvTensorGraph(TensorGraph):
         print('Starting epoch %i' % epoch)
       for ind, (X_b, y_b, w_b, ids_b) in enumerate(
           dataset.iterbatches(
-              self.batch_size,
-              pad_batches=pad_batches,
-              deterministic=deterministic)):
+            self.batch_size,
+            pad_batches=pad_batches,
+            deterministic=deterministic)):
         d = {}
         for index, label in enumerate(self.my_labels):
           if self.mode == 'classification':
@@ -995,7 +802,7 @@ class GraphConvTensorGraph(TensorGraph):
     y_ = []
     for i in range(n_passes):
       generator = self.default_generator(
-          dataset, predict=True, pad_batches=True)
+        dataset, predict=True, pad_batches=True)
       y_.append(self.predict_on_generator(generator, transformers))
 
     # Concatenates along 0-th dimension
